requirements.txt
msgpack_python==0.5.6
numpy==2.1.2
protobuf==5.28.3
